Configure React app for deployment to GitHub Pages

+++ src/App.js
@@ -1,12 +1,14 @@
-import { useEffect, useMemo, useRef, useState } from "react";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import "../node_modules/remixicon/fonts/remixicon.css";
 import "./App.css";
 import TaskItem from "./components/TaskItem";
-
+import { toast } from "react-toastify";
+import Swal from "sweetalert2";
+import MoodToggle from "./components/MoodToggle";
 function App() {
   const [tasks, setTasks] = useState([]);
   const [status, setStatus] = useState("all");
-  console.log("yes");
+
   const inputRef = useRef({
     taskName: "",
     dueDate: "",
@@ -27,24 +29,44 @@ function App() {
 
     inputRef.current.taskName.value = "";
     inputRef.current.dueDate.value = "";
+    toast.success("Task Added Successfully");
   };
 
   // Delete Task
-  const deleteTask = (index) => {
-    const tasksAfterDeletion = [...tasks];
-    tasksAfterDeletion.splice(index, 1);
-    localStorage.setItem("tasks", JSON.stringify(tasksAfterDeletion));
-    setTasks(tasksAfterDeletion);
-  };
+  const deleteTask = useCallback(
+    async (index) => {
+      const result = await Swal.fire({
+        title: "Confirm Deletion",
+        text: "Are you sure you want to delete?",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#d33",
+        cancelButtonColor: "#3085d6",
+        confirmButtonText: "Yes, delete it!",
+      });
+      if (result.isConfirmed) {
+        const tasksAfterDeletion = [...tasks];
+        tasksAfterDeletion.splice(index, 1);
+        localStorage.setItem("tasks", JSON.stringify(tasksAfterDeletion));
+        setTasks(tasksAfterDeletion);
+        toast.success("Task deleted Successfully");
+      }
+    },
+    [tasks]
+  );
 
   // Complete Tasks
-  const completeTask = (index) => {
-    const tasksAfterCompletion = tasks.map((item, i) => {
-      return i === index ? { ...item, isCompleted: true } : item;
-    });
-    setTasks(tasksAfterCompletion);
-    localStorage.setItem("tasks", JSON.stringify(tasksAfterCompletion));
-  };
+  const completeTask = useCallback(
+    (index) => {
+      const tasksAfterCompletion = tasks.map((item, i) => {
+        return i === index ? { ...item, isCompleted: true } : item;
+      });
+      setTasks(tasksAfterCompletion);
+      localStorage.setItem("tasks", JSON.stringify(tasksAfterCompletion));
+      toast.success("Task Completed");
+    },
+    [tasks]
+  );
 
   // Filter tasks
   const filterTasks = useMemo(() => {
@@ -59,7 +81,7 @@ function App() {
       })
     );
   }, [status, tasks]);
-  console.log(filterTasks);
+  
 
   useEffect(() => {
     const tasksStorage = localStorage.getItem("tasks");
@@ -68,11 +90,12 @@ function App() {
     }
   }, []);
   return (
-    <div className="tasks bg-white py-5 px-4 shadow-sm w-75 mt-5 m-auto">
+    <div className="tasks bg-white py-5 px-4 shadow-sm  mt-5 m-auto">
+      <MoodToggle />
       <h1 className="mb-4 text-center fw-bold">Task Tracker</h1>
       <form onSubmit={addTask} className="d-flex flex-column gap-2">
         <div className="mb-3">
-          <label htmlFor="taskName" className="form-label">
+          <label htmlFor="taskName" className="form-label fw-bold">
             Task Name
           </label>
           <input
@@ -85,12 +108,12 @@ function App() {
           />
         </div>
         <div className="mb-3">
-          <label htmlFor="dueDate" className="form-label">
+          <label htmlFor="dueDate" className="form-label fw-bold">
             Due Date
           </label>
           <input
             type="date"
-            className="form-control"
+            className="form-control "
             id="dueDate"
             placeholder="enter a due date"
             ref={(el) => (inputRef.current.dueDate = el)}
@@ -103,29 +126,35 @@ function App() {
         </div>
       </form>
 
-      <div className="filtered-Tasks mb-5 d-flex align-items-center gap-2 py-3 border-2 border-top">
+      <div className="filtered-Tasks mb-5 d-flex align-items-center gap-2 flex-wrap py-3 border-2 border-top">
         <button
           onClick={() => setStatus("all")}
-          className={`btn ${status === "all" ? "bg-success" : "bg-secondary"}  text-white  fw-meduim shadow-sm `}
+          className={`btn  ${
+            status === "all" ? "bg-success" : "bg-secondary"
+          }  text-white fw-normal   fw-lg-lighter shadow-sm `}
         >
           All Tasks
         </button>
         <button
           onClick={() => setStatus("completed")}
-          className={`btn ${status === "completed" ? "bg-success" : "bg-secondary"}  text-white  fw-meduim shadow-sm`}
+          className={`btn ${
+            status === "completed" ? "bg-success" : "bg-secondary"
+          }  text-white  fw-meduim shadow-sm`}
         >
           Completed Tasks
         </button>
         <button
           onClick={() => setStatus("incomplete")}
-          className={`btn ${status === "incomplete" ? "bg-success" : "bg-secondary"}  text-white  fw-meduim shadow-sm`}
+          className={`btn  ${
+            status === "incomplete" ? "bg-success" : "bg-secondary"
+          }  text-white  fw-meduim shadow-sm`}
         >
           Incomplete tasks
         </button>
       </div>
       <div className="list-of-tasks">
         {filterTasks.length > 0 ? (
-          filterTasks.map(({ id,taskName, dueDate, isCompleted }, index) => (
+          filterTasks.map(({ id, taskName, dueDate, isCompleted }, index) => (
             <TaskItem
               key={id}
               taskName={taskName}
@@ -133,6 +162,7 @@ function App() {
               deleteTask={() => deleteTask(index)}
               completeTask={() => completeTask(index)}
               isCompleted={isCompleted}
+              status={status}
             />
           ))
         ) : (
